Modify title and Abstract
The proposed changes provide for emphasis of the fact that ChainLocks strengthens the consensus mechanism.  

The introduction of the term "post Nakamoto consensus" will help me in the literature.  People can use the term ChainLocks when they want to talk about the technology.  I can use post Nakamoto Consensus when referring to how the game theory is different.